Project phase 1: Build catalog of event times to reprocess
This module is part of the Phase 1 in the Repicking (RP) project

The objective is to use the previous catalogs of events and SP ATP from
the student G.R.P. and produce a single catalog containing the file
names and an the mean of the SP for each event.
——
Project successfully finished.
Output file: project_files/catalog_events_n_time.csv