Potential event detection using G.R.P. previous catalog as source
RP_SNR_main: This module is used as a preliminar evaluation of G.R.P.
event catalog.
It evaluates the signals by calculation the ES SNR, then export the
event catalog  with some measured propetrties of each event, that are
strong indicators of event presence

om_ES: Functions for microseismic event detection based on
Energy-Stacking

RP_catalog: This module is part of the Phase 1 in the Repicking (RP)
project

The objective is to use the previous catalogs of events and SP ATP from
the student G.R.P.and produce a single catalog containing the file
names and an the mean of the SP for each event.

om_char_functions: Updated module for calculation of Characteristic
functions for ATP